refactor: Divide response and request into dto
